Added Secutity Features to all the methods and all the controllers

+++ application/controllers/MaintainVendorGST.php
@@ -27,56 +27,60 @@ class MaintainVendorGST extends MY_Controller
 //    function for saving GST Data into database
     public function saveVendorGST()
     {
+        if ($this->authorizeOnly(['user'])) {
 
 
-        $this->form_validation->set_error_delimiters("<p class=text-danger>", "</p>");
-        if ($this->form_validation->run('saveVendorGST') == true) {
+            $this->form_validation->set_error_delimiters("<p class=text-danger>", "</p>");
+            if ($this->form_validation->run('saveVendorGST') == true) {
 
-          $config = array(
-           'upload_path' => "./assets/upload",
-           'allowed_types' => "gif|jpg|png|jpeg",
-           'overwrite' => False,
-           'encrypt_name'=> TRUE,
-            );
+                $config = array(
+                    'upload_path' => "./assets/upload",
+                    'allowed_types' => "gif|jpg|png|jpeg",
+                    'overwrite' => False,
+                    'encrypt_name' => TRUE,
+                );
 
-            $this->load->library('upload',$config);
-            $gstn_image=$this->upload->do_upload('gstn_image');
-             $upload_gstn_image = $this->upload->data('file_name');
-              $picture_gstn_image="assets/upload/".$upload_gstn_image;
-            $gstn_image = array('gstn_image' => $picture_gstn_image);
+                $this->load->library('upload', $config);
+                $gstn_image = $this->upload->do_upload('gstn_image');
+                $upload_gstn_image = $this->upload->data('file_name');
+                $picture_gstn_image = "assets/upload/" . $upload_gstn_image;
+                $gstn_image = array('gstn_image' => $picture_gstn_image);
 
-            //saving Record Here
-            $newGSTNData = $this->input->post();
-            unset($newGSTNData['btnSaveRecord']);
-            $newGSTNData['gstn_image']=$gstn_image['gstn_image'];
+                //saving Record Here
+                $newGSTNData = $this->input->post();
+                unset($newGSTNData['btnSaveRecord']);
+                $newGSTNData['gstn_image'] = $gstn_image['gstn_image'];
 
-            //checking if record for particular state is already present or not
-            $allGSTRecordsofVendor = $this->Dbmodel->getVendorGSTArray($newGSTNData['v_id']);
+                //checking if record for particular state is already present or not
+                $allGSTRecordsofVendor = $this->Dbmodel->getVendorGSTArray($newGSTNData['v_id']);
 
-            $isRecordThere = false;
+                $isRecordThere = false;
 
-            foreach ($allGSTRecordsofVendor as $record) {
-                if ($record['state'] == $newGSTNData['state'])
-                    $isRecordThere = true;
-            }
+                foreach ($allGSTRecordsofVendor as $record) {
+                    if ($record['state'] == $newGSTNData['state'])
+                        $isRecordThere = true;
+                }
 
-            if ($isRecordThere == false) {
-                $this->Dbmodel->addVendorGST($newGSTNData);
-                $this->session->set_flashdata('save', 'Your Record has been Saved Successfully.');
+                if ($isRecordThere == false) {
+                    $this->Dbmodel->addVendorGST($newGSTNData);
+                    $this->session->set_flashdata('save', 'Your Record has been Saved Successfully.');
+
+                } else {
+                    $this->session->set_flashdata('warning', 'Entry for this State is already there for this Vendor.');
+                }
+
+                //-------------
+
+                return redirect('MaintainVendorGST');
+
 
             } else {
-                $this->session->set_flashdata('warning', 'Entry for this State is already there for this Vendor.');
+                $this->session->set_flashdata('warning', 'Please check the fields.');
+                $vendorData['data'] = $this->Dbmodel->getVendorObject(-1);
+                $this->load->view('user/maintain_Vendor_GST.php', $vendorData);
             }
-
-            //-------------
-
-            return redirect('MaintainVendorGST');
-
-
         } else {
-            $this->session->set_flashdata('warning', 'Please check the fields.');
-            $vendorData['data'] = $this->Dbmodel->getVendorObject(-1);
-            $this->load->view('user/maintain_Vendor_GST.php', $vendorData);
+            return redirect('Login');
         }
 
 
@@ -116,7 +120,7 @@ class MaintainVendorGST extends MY_Controller
                    <td>' . $data->state . '</td>
                    <td>' . $data->gstn_no . '</td>
                  
-                   <td><a href=\''.base_url($data->gstn_image).'\'target=\'_blank\'>'.$data->gstn_image .'</a></td>
+                   <td><a href=\'' . base_url($data->gstn_image) . '\'target=\'_blank\'>' . $data->gstn_image . '</a></td>
                    <td><a href=\'MaintainVendorGST\deleteGST\\' . $data->gstn_no . '\' class=\'btn btn-sm btn-danger\' id=>DELETE</a></td>
                </tr>';
                 }
@@ -137,10 +141,14 @@ class MaintainVendorGST extends MY_Controller
 //    Function to delete specific GSTN from the table (passed by parameter)
     public function deleteGST($gst_no)
     {
+        if ($this->authorizeOnly(['user'])) {
 
-        $this->Dbmodel->deleteVendorGSTN($gst_no);
-        $this->session->set_flashdata('info', 'Record has been successfully Deleted.');
-        redirect('MaintainVendorGST');
+            $this->Dbmodel->deleteVendorGSTN($gst_no);
+            $this->session->set_flashdata('info', 'Record has been successfully Deleted.');
+            redirect('MaintainVendorGST');
+        } else {
+            return redirect('Login');
+        }
 
 
     }
